Fix bug in ordering of genotype samples.
The individuals in samples.txt used to subset are not in the same order
as they appear in the genotype files. Thus when using list
comprehensions to subset, only those individuals get saved, but their
column headers are completely random. I solved this by sorting the
specific individuals to subset by their presence in the .fam file.

http://stackoverflow.com/a/23069106/2483477